rename: FollowButton.jsx 파일 위치 변경 @/components/common/button => @/components/celebrity
